Moved old code, changed readmes
