+++ Arduino Library/README.md
@@ -1,6 +1,6 @@
 # Arduino Library
 
-The WidowX Arduino Library was created to simplify the operation of the [WidowX Robot Arm](https://www.trossenrobotics.com/widowxrobotarm) of [Trossen Robotics](https://www.trossenrobotics.com/) when controlled with the [Arbotix-M Robocontroller](https://learn.trossenrobotics.com/arbotix). It is based on the [ax12.h library](https://github.com/vanadiumlabs/arbotix/blob/master/libraries/Bioloid/ax12.h). Also, inside some of its functions, it uses the Matrix object as defined in [BasicLinearAlgebra.h](https://github.com/tomstewart89/BasicLinearAlgebra). Hence, it requires installation in order to work properly.
+The WidowX Arduino Library was created to simplify the operation of the [WidowX Robot Arm](https://www.trossenrobotics.com/widowxrobotarm) of [Trossen Robotics](https://www.trossenrobotics.com/) when controlled with the [Arbotix-M Robocontroller](https://learn.trossenrobotics.com/arbotix). It is based on the [ax12.h library](https://github.com/vanadiumlabs/arbotix/blob/master/libraries/Bioloid/ax12.h). 
 
 For simplicity, each motor's angle is labeled Q1, Q2, Q3, Q4, Q5 and Q6. So whenever you see **Qn**, know that it is reffering to the angle of the motor.
 
@@ -8,11 +8,9 @@ For simplicity, each motor's angle is labeled Q1, Q2, Q3, Q4, Q5 and Q6. So when
 
 First, you’ll need to follow the [ArbotiX-M Robocontroller Getting Started Guide](https://learn.trossenrobotics.com/arbotix/arbotix-quick-start.html) and test the correct connectivity with your WidowX arm. Be mindful of the Arduino version that is being installed, since the documentation specifies that Arduino 1.0.6 is the only one that works with the ArbotiX.
 
-Once you complete this guide, open the Arduino IDE and access to the library manager via **Sketch > Include Library > Manage Libraries** and look for _Basic Linear Algebra_. Select the library by Tom Stewart and install it.
-
 Afterwards, go to the [WidowX Library repo](https://github.com/LeninSG21/WidowX), download it as a zip and extract it. Then, go to the folder ArduinoLibrary and copy the WidowX folder into the Arduino/libraries directory. Open again the Arduino IDE and the WidowX library should appear under the available libraries at **Sketch > Include Library > Contributed Libraries**.
 
-Also, you should see in the same space the Bioloid and BasicLinearAlgebra libraries. That means that everything is installed properly and that the arm is ready to be used.
+Also, you should see in the same space the Bioloid library. That means that everything is installed properly and that the arm is ready to be used.
 
 To confirm the installation, open the file [**Arduino Library > Examples > testArm.ino**](https://github.com/LeninSG21/WidowX/blob/master/Arduino%20Library/Examples/testArm/testArm.ino). The arm should move to rest position, then home, then center and finally rest position again. Open the Serial terminal at 115,200 bps to check the messages from the test.
 
@@ -200,11 +198,10 @@ The next picture shows the motors numbered. Remember that the idx values are fix
 
 #### void moveArmRd(float Px, float Py, float Pz, Matrix<3, 3> &Rd)
 
-> Moves the center of the gripper to the specified coordinates Px, Py and Pz, and with the desired rotation of the coordinate system of the gripper, as seen from the coordinate system {1} of the robot. For example, when Rd is an identity matrix of 3x3, the gripper's rotation will be such that the orientation of the system {1} and the orientation of the gripper's system will be the same. Thus, it is harder to obtain solutions for the IK. It uses getIK_Rd. This function affects Q1, Q2, Q3, Q4, and Q5. It interpolates the step using a cubic interpolation with the default time. If there is no solution for the IK, the arm does not move, and a message is printed into the serial monitor. Rd is a Matrix object as defined by BasicLinearAlgebra.h
-
+> Moves the center of the gripper to the specified coordinates Px, Py and Pz, and with the desired rotation of the coordinate system of the gripper, as seen from the coordinate system {1} of the robot. For example, when Rd is an identity matrix of 3x3, the gripper's rotation will be such that the orientation of the system {1} and the orientation of the gripper's system will be the same. Thus, it is harder to obtain solutions for the IK. It uses getIK_Rd. This function affects Q1, Q2, Q3, Q4, and Q5. It interpolates the step using a cubic interpolation with the default time. If there is no solution for the IK, the arm does not move, and a message is printed into the serial monitor. Rd is a Matrix given as a bidimensial float array
 #### void moveArmRd(float Px, float Py, float Pz, Matrix<3, 3> &Rd, int time);
 
-> Moves the center of the gripper to the specified coordinates Px, Py and Pz, as seen from the base of the robot, and with the desired rotation of the coordinate system of the gripper, as seen from the coordinate system {1} of the robot. For example, when Rd is an identity matrix of 3x3, the gripper's rotation will be such that the orientation of the system {1} and the orientation of the gripper's system will be the same. Thus, it is harder to obtain solutions for the IK. It uses getIK_Rd. This function affects Q1, Q2, Q3, Q4, and Q5. It interpolates the step using a cubic interpolation with the given time in milliseconds. If there is no solution for the IK, the arm does not move, and a message is printed into the serial monitor. Rd is a Matrix object as defined by BasicLinearAlgebra.h
+> Moves the center of the gripper to the specified coordinates Px, Py and Pz, as seen from the base of the robot, and with the desired rotation of the coordinate system of the gripper, as seen from the coordinate system {1} of the robot. For example, when Rd is an identity matrix of 3x3, the gripper's rotation will be such that the orientation of the system {1} and the orientation of the gripper's system will be the same. Thus, it is harder to obtain solutions for the IK. It uses getIK_Rd. This function affects Q1, Q2, Q3, Q4, and Q5. It interpolates the step using a cubic interpolation with the given time in milliseconds. If there is no solution for the IK, the arm does not move, and a message is printed into the serial monitor. Rd is a Matrix given as a bidimensial float array
 
 #### void moveArmRdBase(float Px, float Py, float Pz, Matrix<3, 3> &RdBase);
