+++ old/WidowX/Examples/testArm/testArm.ino
@@ -1,6 +1,5 @@
 #include <WidowX.h>
 #include <poses.h>
-#include <ax12.h>
 
 WidowX widow = WidowX();
